+++ test/o_security_token.js
@@ -518,21 +518,21 @@ contract('SecurityToken', accounts => {
         });
 
         it("Should get the modules of the securityToken by name", async () => {
-            let moduleData = await I_SecurityToken.getModuleByName.call(stoKey, "CappedSTO");
-            assert.equal(web3.utils.toAscii(moduleData[0]).replace(/\u0000/g, ''), "CappedSTO");
-            assert.equal(moduleData[1], I_CappedSTO.address);
+            let moduleData = await I_SecurityToken.getAllModulesByName.call(stoKey, "CappedSTO");
+            assert.equal(web3.utils.toAscii(moduleData[0][0]).replace(/\u0000/g, ''), "CappedSTO");
+            assert.equal(moduleData[1][0], I_CappedSTO.address);
         });
 
         it("Should get the modules of the securityToken by name (not added into the security token yet)", async () => {
-            let moduleData = await I_SecurityToken.getModuleByName.call(permissionManagerKey, "GeneralPermissionManager");
-            assert.equal(web3.utils.toAscii(moduleData[0]).replace(/\u0000/g, ''), "");
-            assert.equal(moduleData[1], "0x0000000000000000000000000000000000000000");
+            let moduleData = await I_SecurityToken.getAllModulesByName.call(permissionManagerKey, "GeneralPermissionManager");
+            assert.equal(moduleData[0].length, 0);
+            assert.equal(moduleData[1].length, 0);
         });
 
         it("Should get the modules of the securityToken by name (not added into the security token yet)", async () => {
-            let moduleData = await I_SecurityToken.getModuleByName.call(transferManagerKey, "CountTransferManager");
-            assert.equal(web3.utils.toAscii(moduleData[0]).replace(/\u0000/g, ''), "");
-            assert.equal(moduleData[1], "0x0000000000000000000000000000000000000000");
+            let moduleData = await I_SecurityToken.getAllModulesByName.call(transferManagerKey, "CountTransferManager");
+            assert.equal(moduleData[0].length, 0);
+            assert.equal(moduleData[1].length, 0);
         });
 
         it("Should fail in updating the token details", async() => {